+++ src/room_config.h
@@ -19,43 +19,43 @@ public:
 	bool ShowBorder;
 	bool ShowOuterWalls;
 	// room noise
-	double IsoValue;
-	double Curve;
-	double Tilt;
-	double FalloffAboveCeiling;
+	float IsoValue;
+	float Curve;
+	float Tilt;
+	float FalloffAboveCeiling;
 	bool Interpolate;
 	bool RemoveOrphans;
 	// room border
 	bool UseBorderNoise;
 	int BorderSize;
-	double SmoothBorderNoise;
+	float SmoothBorderNoise;
 	int FalloffNearBorder;
 
 	bool GetShowNoise();
 	bool GetShowBorder();
 	bool GetShowOuterWalls();
-	double GetIsoValue();
-	double GetCurve();
-	double GetTilt();
-	double GetFalloffAboveCeiling();
+	float GetIsoValue();
+	float GetCurve();
+	float GetTilt();
+	float GetFalloffAboveCeiling();
 	bool GetInterpolate();
 	bool GetRemoveOrphans();
 	bool GetUseBorderNoise();
 	int GetBorderSize();
-	double GetSmoothBorderNoise();
+	float GetSmoothBorderNoise();
 	int GetFalloffNearBorder();
 
 	void SetShowNoise(bool p_ShowNoise);
 	void SetShowBorder(bool p_ShowBorder);
 	void SetShowOuterWalls(bool p_ShowOuterWalls);
-	void SetIsoValue(double p_IsoValue);
-	void SetCurve(double p_Curve);
-	void SetTilt(double p_Tilt);
-	void SetFalloffAboveCeiling(double p_FalloffAboveCeiling);
+	void SetIsoValue(float p_IsoValue);
+	void SetCurve(float p_Curve);
+	void SetTilt(float p_Tilt);
+	void SetFalloffAboveCeiling(float p_FalloffAboveCeiling);
 	void SetInterpolate(bool p_Interpolate);
 	void SetRemoveOrphans(bool p_RemoveOrphans);
 	void SetUseBorderNoise(bool p_UseBorderNoise);
 	void SetBorderSize(int p_BorderSize);
-	void SetSmoothBorderNoise(double p_SmoothBorderNoise);
+	void SetSmoothBorderNoise(float p_SmoothBorderNoise);
 	void SetFalloffNearBorder(int p_FalloffNearBorder);
 };
